requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.20.0
huggingface-hub>=0.16.0

# Quantization dependencies
bitsandbytes>=0.41.0
llmcompressor>=0.1.0

# Training and evaluation
scikit-learn>=1.3.0
numpy>=1.24.0
tensorboard>=2.13.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0

# Optional: For development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
